requirements.txt
app~=0.0.1
click==7.1.2
Flask==1.1.2
Jinja2==2.11.3
pymongo~=4.0.1
python-dotenv==0.19.0
requests~=2.27.1
Werkzeug==1.0.1